+++ mediation/Fyber/source/plugin/Assets/GoogleMobileAds/Common/Mediation/Fyber/IFyberClient.cs
@@ -19,5 +19,8 @@ namespace GoogleMobileAds.Common.Mediation.Fyber
         void SetGDPRConsent(bool consent);
         void SetGDPRConsentString(string consentString);
         void ClearGDPRConsentData();
+
+        void SetCCPAString(string ccpaString);
+        void ClearCCPAString();
     }
 }